Adding new pages to the app
I added new pages to the app. The new pages that I aded were, About, Contacts, and FavoritesPage, rn they are on the navigation bar upon clicking them takes the user to the respective pages. Rn they also do not have any CSS added so they are not aligned properly.